Crate rustc_errors

Unstable (rustc_private)

Reexports

pub use emitter::ColorConfig;

pub use diagnostic::Diagnostic;

pub use diagnostic::SubDiagnostic;

pub use diagnostic_builder::DiagnosticBuilder;

Modules

diagnostic 	 [
Unstable
]

diagnostic_builder 	 [
Unstable
]

emitter 	 [
Unstable
]

registry 	 [
Unstable
]

snippet 	 [
Unstable
]

styled_buffer 	 [
Unstable
]

Structs

CodeSuggestion 	 [
Unstable
]

ExplicitBug 	 [
Unstable
]
Signifies that the compiler died with an explicit call to .bug or .span_bug rather than a failed assertion, etc.

FatalError 	 [
Unstable
]
Used as a return value to signify a fatal error occurred. (It is also used as the argument to panic at the moment, but that will eventually not be true.)

Handler 	 [
Unstable
]
A handler deals with errors; certain errors (fatal, bug, unimpl) may cause immediate exit, others log errors for later reporting.

Enums

Level 	 [
Unstable
]

RenderSpan 	 [
Unstable
]

Traits

CodeMapper 	 [
Unstable
]

Functions

expect 	 [
Unstable
]