Struct rustc_errors::annotate_snippet_emitter_writer::AnnotateSnippetEmitterWriter

pub struct AnnotateSnippetEmitterWriter {
    source_map: Option<Lrc<SourceMap>>,
    fluent_bundle: Option<Lrc<FluentBundle>>,
    fallback_bundle: LazyFallbackBundle,
    short_message: bool,
    ui_testing: bool,
    macro_backtrace: bool,
}

Generates diagnostics using annotate-snippet

Fields
source_map: Option<Lrc<SourceMap>>fluent_bundle: Option<Lrc<FluentBundle>>fallback_bundle: LazyFallbackBundleshort_message: bool

If true, hides the longer explanation text

ui_testing: bool

If true, will normalize line numbers with LL to prevent noise in UI test diffs.

macro_backtrace: bool
Implementations

impl AnnotateSnippetEmitterWriter

pub fn new(
source_map: Option<Lrc<SourceMap>>,
fluent_bundle: Option<Lrc<FluentBundle>>,
fallback_bundle: LazyFallbackBundle,
short_message: bool,
macro_backtrace: bool
) -> Self

pub fn ui_testing(self, ui_testing: bool) -> Self

Allows to modify Self to enable or disable the ui_testing flag.

If this is set to true, line numbers will be normalized as LL in the output.

fn emit_messages_default(
&mut self,
level: &Level,
messages: &[(DiagnosticMessage, Style)],
args: &FluentArgs<'_>,
code: &Option<DiagnosticId>,
msp: &MultiSpan,
_children: &[SubDiagnostic],
_suggestions: &[CodeSuggestion]
)