Struct rustc_errors::diagnostic_builder::DiagnosticBuilder

#[must_use]

pub struct DiagnosticBuilder<'a> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

Used for emitting structured error messages and other diagnostic information.

Methods

impl<'a> DiagnosticBuilder<'a>

fn emit(&mut self)

🔬 This is a nightly-only experimental API. (rustc_private)

Emit the diagnostic.

fn span_label(&mut self, span: Span, label: &Display) -> &mut Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn note_expected_found(&mut self,
label: &Display,
expected: &Display,
found: &Display)
-> &mut Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn note_expected_found_extra(&mut self,
label: &Display,
expected: &Display,
found: &Display,
expected_extra: &Display,
found_extra: &Display)
-> &mut Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn note(&mut self, msg: &str) -> &mut Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn span_note<S: Into<MultiSpan>>(&mut self, sp: S, msg: &str) -> &mut Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn warn(&mut self, msg: &str) -> &mut Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn span_warn<S: Into<MultiSpan>>(&mut self, sp: S, msg: &str) -> &mut Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn help(&mut self, msg: &str) -> &mut Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn span_help<S: Into<MultiSpan>>(&mut self, sp: S, msg: &str) -> &mut Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn span_suggestion<S: Into<MultiSpan>>(&mut self,
sp: S,
msg: &str,
suggestion: String)
-> &mut Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn set_span<S: Into<MultiSpan>>(&mut self, sp: S) -> &mut Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn code(&mut self, s: String) -> &mut Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn new(handler: &'a Handler,
level: Level,
message: &str)
-> DiagnosticBuilder<'a>

🔬 This is a nightly-only experimental API. (rustc_private)

Convenience function for internal use, clients should use one of the struct_* methods on Handler.

fn new_with_code(handler: &'a Handler,
level: Level,
code: Option<String>,
message: &str)
-> DiagnosticBuilder<'a>

🔬 This is a nightly-only experimental API. (rustc_private)

Convenience function for internal use, clients should use one of the struct_* methods on Handler.

fn into_diagnostic(self) -> Diagnostic

🔬 This is a nightly-only experimental API. (rustc_private)