Enum rustc_errors::emitter::Destination

pub enum Destination {
    Terminal(Box<StderrTerminal>),
    BufferedTerminal(Box<BufferedStderr>),
    Raw(Box<Write + Send>),
}

Variants
Terminal(Box<StderrTerminal>)BufferedTerminal(Box<BufferedStderr>)Raw(Box<Write + Send>)
Methods

impl Destination

Important traits for Destination
impl Write for Destination

fn from_stderr() -> Destination

When not on Windows, prefer the buffered terminal so that we can buffer an entire error to be emitted at one time.

fn apply_style(&mut self, lvl: Level, style: Style) -> Result<()>

fn start_attr(&mut self, attr: Attr) -> Result<()>

fn reset_attrs(&mut self) -> Result<()>