Struct rustc_errors::emitter::BufferedWriter

pub struct BufferedWriter { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Buffered writer gives us a way on Unix to buffer up an entire error message before we output it. This helps to prevent interleaving of multiple error messages when multiple compiler processes error simultaneously