Struct rustc_errors::emitter::EmitterWriter

pub struct EmitterWriter {
    dst: Destination,
    cm: Option<Rc<CodeMapper>>,
    short_message: bool,
}

Fields
dst: Destination cm: Option<Rc<CodeMapper>> short_message: bool
Methods

impl EmitterWriter

pub fn stderr(
color_config: ColorConfig,
code_map: Option<Rc<CodeMapper>>,
short_message: bool
) -> EmitterWriter

pub fn new(
dst: Box<Write + Send>,
code_map: Option<Rc<CodeMapper>>,
short_message: bool
) -> EmitterWriter

fn preprocess_annotations(
&mut self,
msp: &MultiSpan
) -> Vec<FileWithAnnotatedLines>

fn render_source_line(
&self,
buffer: &mut StyledBuffer,
file: Rc<FileMap>,
line: &Line,
width_offset: usize,
code_offset: usize
) -> Vec<(usize, Style)>

fn get_multispan_max_line_num(&mut self, msp: &MultiSpan) -> usize

fn get_max_line_num(
&mut self,
span: &MultiSpan,
children: &Vec<SubDiagnostic>
) -> usize

fn fix_multispan_in_std_macros(
&mut self,
span: &mut MultiSpan,
always_backtrace: bool
) -> bool

fn fix_multispans_in_std_macros(
&mut self,
span: &mut MultiSpan,
children: &mut Vec<SubDiagnostic>,
backtrace: bool
)

fn msg_to_buffer(
&self,
buffer: &mut StyledBuffer,
msg: &[(String, Style)],
padding: usize,
label: &str,
override_style: Option<Style>
)

Add a left margin to every line but the first, given a padding length and the label being displayed, keeping the provided highlighting.

fn emit_message_default(
&mut self,
msp: &MultiSpan,
msg: &Vec<(String, Style)>,
code: &Option<DiagnosticId>,
level: &Level,
max_line_num_len: usize,
is_secondary: bool
) -> Result<()>

fn emit_suggestion_default(
&mut self,
suggestion: &CodeSuggestion,
level: &Level,
max_line_num_len: usize
) -> Result<()>

fn emit_messages_default(
&mut self,
level: &Level,
message: &Vec<(String, Style)>,
code: &Option<DiagnosticId>,
span: &MultiSpan,
children: &Vec<SubDiagnostic>,
suggestions: &[CodeSuggestion]
)