Struct rustc_errors::emitter::HumanEmitter

pub struct HumanEmitter {
    dst: IntoDynSyncSend<Destination>,
    sm: Option<Lrc<SourceMap>>,
    fluent_bundle: Option<Lrc<FluentBundle>>,
    fallback_bundle: LazyFallbackBundle,
    short_message: bool,
    teach: bool,
    ui_testing: bool,
    ignored_directories_in_source_blocks: Vec<String>,
    diagnostic_width: Option<usize>,
    macro_backtrace: bool,
    track_diagnostics: bool,
    terminal_url: TerminalUrl,
}

Handles the writing of HumanReadableErrorType::Default and HumanReadableErrorType::Short

Fields
dst: IntoDynSyncSend<Destination>sm: Option<Lrc<SourceMap>>fluent_bundle: Option<Lrc<FluentBundle>>fallback_bundle: LazyFallbackBundleshort_message: boolteach: boolui_testing: boolignored_directories_in_source_blocks: Vec<String>diagnostic_width: Option<usize>macro_backtrace: booltrack_diagnostics: boolterminal_url: TerminalUrl
Implementations

impl HumanEmitter

pub fn sm(self, value: Option<Lrc<SourceMap>>) -> Self

pub fn fluent_bundle(self, value: Option<Lrc<FluentBundle>>) -> Self

pub fn short_message(self, value: bool) -> Self

pub fn teach(self, value: bool) -> Self

pub fn ui_testing(self, value: bool) -> Self

pub fn ignored_directories_in_source_blocks(self, value: Vec<String>) -> Self

pub fn diagnostic_width(self, value: Option<usize>) -> Self

pub fn macro_backtrace(self, value: bool) -> Self

pub fn track_diagnostics(self, value: bool) -> Self

pub fn terminal_url(self, value: TerminalUrl) -> Self

impl HumanEmitter

pub fn new( dst: Destination, fallback_bundle: LazyFallbackBundle, ) -> HumanEmitter

fn maybe_anonymized(&self, line_num: usize) -> Cow<'static, str>

fn draw_line( &self, buffer: &mut StyledBuffer, source_string: &str, line_index: usize, line_offset: usize, width_offset: usize, code_offset: usize, margin: Margin, )

fn render_source_line( &self, buffer: &mut StyledBuffer, file: Lrc<SourceFile>, line: &Line, width_offset: usize, code_offset: usize, margin: Margin, ) -> Vec<(usize, Style)>

fn get_multispan_max_line_num(&mut self, msp: &MultiSpan) -> usize

fn get_max_line_num(&mut self, span: &MultiSpan, children: &[Subdiag]) -> usize

fn msgs_to_buffer( &self, buffer: &mut StyledBuffer, msgs: &[(DiagMessage, Style)], args: &FluentArgs<'_>, padding: usize, label: &str, override_style: Option<Style>, )

Adds a left margin to every line but the first, given a padding length and the label being displayed, keeping the provided highlighting.

fn emit_messages_default_inner( &mut self, msp: &MultiSpan, msgs: &[(DiagMessage, Style)], args: &FluentArgs<'_>, code: &Option<ErrCode>, level: &Level, max_line_num_len: usize, is_secondary: bool, emitted_at: Option<&DiagLocation>, ) -> Result<()>

fn emit_suggestion_default( &mut self, span: &MultiSpan, suggestion: &CodeSuggestion, args: &FluentArgs<'_>, level: &Level, max_line_num_len: usize, ) -> Result<()>

fn emit_messages_default( &mut self, level: &Level, messages: &[(DiagMessage, Style)], args: &FluentArgs<'_>, code: &Option<ErrCode>, span: &MultiSpan, children: &[Subdiag], suggestions: &[CodeSuggestion], emitted_at: Option<&DiagLocation>, )

fn draw_code_line( &self, buffer: &mut StyledBuffer, row_num: &mut usize, highlight_parts: &[SubstitutionHighlight], line_num: usize, line_to_add: &str, show_code_change: DisplaySuggestion, max_line_num_len: usize, file_lines: &FileLines, is_multiline: bool, )