Struct rustc_errors::emitter::SilentEmitter

pub struct SilentEmitter {
    pub fatal_handler: Handler,
    pub fatal_note: Option<String>,
}

An emitter that does nothing when emitting a non-fatal diagnostic. Fatal diagnostics are forwarded to fatal_handler to avoid silent failures of rustc, as witnessed e.g. in issue #89358.

Fields
fatal_handler: Handlerfatal_note: Option<String>