Struct rustc_errors::json::DiagnosticSpan

struct DiagnosticSpan {

    file_name: String,
    byte_start: u32,
    byte_end: u32,
    line_start: usize,
    line_end: usize,
    column_start: usize,
    column_end: usize,
    is_primary: bool,
    text: Vec<DiagnosticSpanLine>,
    label: Option<String>,
    suggested_replacement: Option<String>,
    suggestion_applicability: Option<Applicability>,
    expansion: Option<Box<DiagnosticSpanMacroExpansion>>,

}

Fields
file_name: Stringbyte_start: u32byte_end: u32line_start: usize

1-based.

line_end: usizecolumn_start: usize

1-based, character offset.

column_end: usizeis_primary: bool

Is this a “primary” span – meaning the point, or one of the points, where the error occurred?

text: Vec<DiagnosticSpanLine>

Source text from the start of line_start to the end of line_end.

label: Option<String>

Label that should be placed at this location (if any)

suggested_replacement: Option<String>

If we are suggesting a replacement, this will contain text that should be sliced in atop this span.

suggestion_applicability: Option<Applicability>

If the suggestion is approximate

expansion: Option<Box<DiagnosticSpanMacroExpansion>>

Macro invocations that created the code at this span, if any.

Implementations

impl DiagnosticSpan

fn from_span_label(
span: SpanLabel,
suggestion: Option<(&String, Applicability)>,
args: &FluentArgs<'_>,
je: &JsonEmitter
) -> DiagnosticSpan

fn from_span_etc(
span: Span,
is_primary: bool,
label: Option<String>,
suggestion: Option<(&String, Applicability)>,
je: &JsonEmitter
) -> DiagnosticSpan

fn from_span_full(
span: Span,
is_primary: bool,
label: Option<String>,
suggestion: Option<(&String, Applicability)>,
backtrace: impl Iterator<Item = ExpnData>,
je: &JsonEmitter
) -> DiagnosticSpan

fn from_multispan(
msp: &MultiSpan,
args: &FluentArgs<'_>,
je: &JsonEmitter
) -> Vec<DiagnosticSpan>

fn from_suggestion(
suggestion: &CodeSuggestion,
args: &FluentArgs<'_>,
je: &JsonEmitter
) -> Vec<DiagnosticSpan>