Enum rustc_errors::snippet::AnnotationType

pub enum AnnotationType {
    Singleline,
    Minimized,
    Multiline(MultilineAnnotation),
    MultilineStart(usize),
    MultilineEnd(usize),
    MultilineLine(usize),
}

🔬 This is a nightly-only experimental API. (rustc_private)

Variants
Singleline

🔬 This is a nightly-only experimental API. (rustc_private)

Annotation under a single line of code

Minimized

🔬 This is a nightly-only experimental API. (rustc_private)

Annotation under the first character of a multiline span

Multiline(MultilineAnnotation)

🔬 This is a nightly-only experimental API. (rustc_private)

Annotation enclosing the first and last character of a multiline span

MultilineStart(usize)

🔬 This is a nightly-only experimental API. (rustc_private)

Annotation marking the first character of a fully shown multiline span

MultilineEnd(usize)

🔬 This is a nightly-only experimental API. (rustc_private)

Annotation marking the last character of a fully shown multiline span

MultilineLine(usize)

🔬 This is a nightly-only experimental API. (rustc_private)

Line at the left enclosing the lines of a fully shown multiline span