Struct rustc_errors::CodeSuggestion

pub struct CodeSuggestion {
    pub substitutions: Vec<Substitution>,
    pub msg: String,
    pub show_code_when_inline: bool,
    pub applicability: Applicability,
}

Fields
substitutions: Vec<Substitution>

Each substitute can have multiple variants due to multiple applicable suggestions

foo.bar might be replaced with a.b or x.y by replacing foo and bar on their own:

vec![
    Substitution { parts: vec![(0..3, "a"), (4..7, "b")] },
    Substitution { parts: vec![(0..3, "x"), (4..7, "y")] },
]

or by replacing the entire span:

vec![
    Substitution { parts: vec![(0..7, "a.b")] },
    Substitution { parts: vec![(0..7, "x.y")] },
]

msg: String show_code_when_inline: bool applicability: Applicability

Whether or not the suggestion is approximate

Sometimes we may show suggestions with placeholders, which are useful for users but not useful for tools like rustfix

Methods

impl CodeSuggestion

pub fn splice_lines(
&self,
cm: &CodeMapperDyn
) -> Vec<(String, Vec<SubstitutionPart>)>

Returns the assembled code suggestions and whether they should be shown with an underline.