Struct rustc_errors::Diagnostic

#[must_use]

pub struct Diagnostic {
    pub level: Level,
    pub message: Vec<(String, Style)>,
    pub code: Option<DiagnosticId>,
    pub span: MultiSpan,
    pub children: Vec<SubDiagnostic>,
    pub suggestions: Vec<CodeSuggestion>,
}

Fields
level: Level message: Vec<(String, Style)> code: Option<DiagnosticId> span: MultiSpan children: Vec<SubDiagnostic> suggestions: Vec<CodeSuggestion>
Methods

impl Diagnostic

pub fn new(level: Level, message: &str) -> Self

pub fn new_with_code(
level: Level,
code: Option<DiagnosticId>,
message: &str
) -> Self

pub fn cancel(&mut self)

Cancel the diagnostic (a structured diagnostic must either be emitted or canceled or it will panic when dropped).

pub fn cancelled(&self) -> bool

pub fn span_label<T: Into<String>>(&mut self, span: Span, label: T) -> &mut Self

Add a span/label to be included in the resulting snippet. This is pushed onto the MultiSpan that was created when the diagnostic was first built. If you don't call this function at all, and you just supplied a Span to create the diagnostic, then the snippet will just include that Span, which is called the primary span.

pub fn note_expected_found(
&mut self,
label: &Display,
expected: DiagnosticStyledString,
found: DiagnosticStyledString
) -> &mut Self

pub fn note_expected_found_extra(
&mut self,
label: &Display,
expected: DiagnosticStyledString,
found: DiagnosticStyledString,
expected_extra: &Display,
found_extra: &Display
) -> &mut Self

pub fn note_trait_signature(
&mut self,
name: String,
signature: String
) -> &mut Self

pub fn note(&mut self, msg: &str) -> &mut Self

pub fn highlighted_note(&mut self, msg: Vec<(String, Style)>) -> &mut Self

pub fn span_note<S: Into<MultiSpan>>(&mut self, sp: S, msg: &str) -> &mut Self

pub fn warn(&mut self, msg: &str) -> &mut Self

pub fn span_warn<S: Into<MultiSpan>>(&mut self, sp: S, msg: &str) -> &mut Self

pub fn help(&mut self, msg: &str) -> &mut Self

pub fn span_help<S: Into<MultiSpan>>(&mut self, sp: S, msg: &str) -> &mut Self

pub fn span_suggestion_short(
&mut self,
sp: Span,
msg: &str,
suggestion: String
) -> &mut Self

Prints out a message with a suggested edit of the code. If the suggestion is presented inline it will only show the text message and not the text.

See CodeSuggestion for more information.

pub fn span_suggestion(
&mut self,
sp: Span,
msg: &str,
suggestion: String
) -> &mut Self

Prints out a message with a suggested edit of the code.

In case of short messages and a simple suggestion, rustc displays it as a label like

"try adding parentheses: (tup.0).1"

The message

should not end in any punctuation (a : is added automatically)

should not be a question

should not contain any parts like "the following", "as shown"

may look like "to do xyz, use" or "to do xyz, use abc"

may contain a name of a function, variable or type, but not whole expressions

See CodeSuggestion for more information.

pub fn multipart_suggestion(
&mut self,
msg: &str,
suggestion: Vec<(Span, String)>
) -> &mut Self

pub fn span_suggestions(
&mut self,
sp: Span,
msg: &str,
suggestions: Vec<String>
) -> &mut Self

Prints out a message with multiple suggested edits of the code.

pub fn span_suggestion_with_applicability(
&mut self,
sp: Span,
msg: &str,
suggestion: String,
applicability: Applicability
) -> &mut Self

This is a suggestion that may contain mistakes or fillers and should be read and understood by a human.

pub fn span_suggestions_with_applicability(
&mut self,
sp: Span,
msg: &str,
suggestions: Vec<String>,
applicability: Applicability
) -> &mut Self

pub fn span_suggestion_short_with_applicability(
&mut self,
sp: Span,
msg: &str,
suggestion: String,
applicability: Applicability
) -> &mut Self

pub fn set_span<S: Into<MultiSpan>>(&mut self, sp: S) -> &mut Self

pub fn code(&mut self, s: DiagnosticId) -> &mut Self

pub fn get_code(&self) -> Option<DiagnosticId>

pub fn message(&self) -> String

pub fn styled_message(&self) -> &Vec<(String, Style)>

pub fn copy_details_not_message(&mut self, from: &Diagnostic)

Used by a lint. Copies over all details but the "main message".