Struct rustc_errors::DiagnosticBuilder

pub struct DiagnosticBuilder<'a, G: EmissionGuarantee> {
    inner: DiagnosticBuilderInner<'a>,
    _marker: PhantomData<G>,
}

Used for emitting structured error messages and other diagnostic information.

If there is some state in a downstream crate you would like to access in the methods of DiagnosticBuilder here, consider extending HandlerFlags, accessed via self.handler.flags.

Fields
inner: DiagnosticBuilderInner<'a>_marker: PhantomData<G>
Implementations

impl<'a> DiagnosticBuilder<'a, ErrorGuaranteed>

pub(crate) fn new_guaranteeing_error<M: Into<DiagnosticMessage>, const L: Level>(
handler: &'a Handler,
message: M
) -> Selfwhere
(): IsError<L>,

Convenience function for internal use, clients should use one of the struct_* methods on Handler.

pub fn forget_guarantee(self) -> DiagnosticBuilder<'a, ()>

Discard the guarantee .emit() would return, in favor of having the type DiagnosticBuilder<'a, ()>. This may be necessary whenever there is a common codepath handling both errors and warnings.

impl<'a> DiagnosticBuilder<'a, ()>

pub(crate) fn new<M: Into<DiagnosticMessage>>(
handler: &'a Handler,
level: Level,
message: M
) -> Self

Convenience function for internal use, clients should use one of the struct_* methods on Handler.

pub(crate) fn new_diagnostic(
handler: &'a Handler,
diagnostic: Diagnostic
) -> Self

Creates a new DiagnosticBuilder with an already constructed diagnostic.

impl<'a> DiagnosticBuilder<'a, Noted>

pub(crate) fn new_note(
handler: &'a Handler,
message: impl Into<DiagnosticMessage>
) -> Self

Convenience function for internal use, clients should use one of the struct_* methods on Handler.

pub(crate) fn new_diagnostic_note(
handler: &'a Handler,
diagnostic: Diagnostic
) -> Self

Creates a new DiagnosticBuilder with an already constructed diagnostic.

impl<'a> DiagnosticBuilder<'a, !>

pub(crate) fn new_fatal(
handler: &'a Handler,
message: impl Into<DiagnosticMessage>
) -> Self

Convenience function for internal use, clients should use one of the struct_* methods on Handler.

pub(crate) fn new_diagnostic_fatal(
handler: &'a Handler,
diagnostic: Diagnostic
) -> Self

Creates a new DiagnosticBuilder with an already constructed diagnostic.

impl<'a, G: EmissionGuarantee> DiagnosticBuilder<'a, G>

pub fn emit(&mut self) -> G

Emit the diagnostic.

pub fn emit_unless(&mut self, delay: bool) -> G

Emit the diagnostic unless delay is true, in which case the emission will be delayed as a bug.

See emit and delay_as_bug for details.

pub fn cancel(self)

Cancel the diagnostic (a structured diagnostic must either be emitted or cancelled or it will panic when dropped).

This method takes self by-value to disallow calling .emit() on it, which may be expected to guarantee the emission of an error, either at the time of the call, or through a prior .emit() call.

pub fn stash(self, span: Span, key: StashKey)

Stashes diagnostic for possible later improvement in a different, later stage of the compiler. The diagnostic can be accessed with the provided span and key through Handler::steal_diagnostic().

As with buffer, this is unless the handler has disabled such buffering.

pub fn into_diagnostic(self) -> Option<(Diagnostic, &'a Handler)>

Converts the builder to a Diagnostic for later emission, unless handler has disabled such buffering, or .emit() was called.

pub fn buffer(self, buffered_diagnostics: &mut Vec<Diagnostic>)

Buffers the diagnostic for later emission, unless handler has disabled such buffering.

pub fn delay_as_bug(&mut self) -> G

Delay emission of this diagnostic as a bug.

This can be useful in contexts where an error indicates a bug but typically this only happens when other compilation errors have already happened. In those cases this can be used to defer emission of this diagnostic as a bug in the compiler only if no other errors have been emitted.

In the meantime, though, callsites are required to deal with the “bug” locally in whichever way makes the most sense.

pub fn downgrade_to_delayed_bug(&mut self) -> &mut Self

See Diagnostic::downgrade_to_delayed_bug().

pub fn span_label(
&mut self,
span: Span,
label: impl Into<SubdiagnosticMessage>
) -> &mut Self

Appends a labeled span to the diagnostic.

Labels are used to convey additional context for the diagnostic’s primary span. They will be shown together with the original diagnostic’s span, not with spans added by span_note, span_help, etc. Therefore, if the primary span is not displayable (because the span is DUMMY_SP or the source code isn’t found), labels will not be displayed either.

Implementation-wise, the label span is pushed onto the MultiSpan that was created when the diagnostic was constructed. However, the label span is not considered a “primary span”; only the Span supplied when creating the diagnostic is primary. See Diagnostic::span_label().

pub fn span_labels(
&mut self,
spans: impl IntoIterator<Item = Span>,
label: impl AsRef<str>
) -> &mut Self

Labels all the given spans with the provided label. See Diagnostic::span_label() for more information. See Diagnostic::span_labels().

pub fn note_expected_found(
&mut self,
expected_label: &dyn Display,
expected: DiagnosticStyledString,
found_label: &dyn Display,
found: DiagnosticStyledString
) -> &mut Self

See Diagnostic::note_expected_found().

pub fn note_expected_found_extra(
&mut self,
expected_label: &dyn Display,
expected: DiagnosticStyledString,
found_label: &dyn Display,
found: DiagnosticStyledString,
expected_extra: &dyn Display,
found_extra: &dyn Display
) -> &mut Self

See Diagnostic::note_expected_found_extra().

pub fn note_unsuccessful_coercion(
&mut self,
expected: DiagnosticStyledString,
found: DiagnosticStyledString
) -> &mut Self

See Diagnostic::note_unsuccessful_coercion().

pub fn note(&mut self, msg: impl Into<SubdiagnosticMessage>) -> &mut Self

See Diagnostic::note().

pub fn note_once(&mut self, msg: impl Into<SubdiagnosticMessage>) -> &mut Self

See Diagnostic::note_once().

pub fn span_note(
&mut self,
sp: impl Into<MultiSpan>,
msg: impl Into<SubdiagnosticMessage>
) -> &mut Self

See Diagnostic::span_note().

pub fn span_note_once(
&mut self,
sp: impl Into<MultiSpan>,
msg: impl Into<SubdiagnosticMessage>
) -> &mut Self

See Diagnostic::span_note_once().

pub fn warn(&mut self, msg: impl Into<SubdiagnosticMessage>) -> &mut Self

See Diagnostic::warn().

pub fn span_warn(
&mut self,
sp: impl Into<MultiSpan>,
msg: impl Into<SubdiagnosticMessage>
) -> &mut Self

See Diagnostic::span_warn().

pub fn help(&mut self, msg: impl Into<SubdiagnosticMessage>) -> &mut Self

See Diagnostic::help().

pub fn span_help(
&mut self,
sp: impl Into<MultiSpan>,
msg: impl Into<SubdiagnosticMessage>
) -> &mut Self

See Diagnostic::span_help().

pub fn help_use_latest_edition(&mut self) -> &mut Self

See Diagnostic::help_use_latest_edition().

pub fn set_is_lint(&mut self) -> &mut Self

See Diagnostic::set_is_lint().

pub fn disable_suggestions(&mut self) -> &mut Self

See Diagnostic::disable_suggestions().

pub fn clear_suggestions(&mut self) -> &mut Self

See Diagnostic::clear_suggestions().

pub fn multipart_suggestion(
&mut self,
msg: impl Into<SubdiagnosticMessage>,
suggestion: Vec<(Span, String)>,
applicability: Applicability
) -> &mut Self

See Diagnostic::multipart_suggestion().

pub fn multipart_suggestion_verbose(
&mut self,
msg: impl Into<SubdiagnosticMessage>,
suggestion: Vec<(Span, String)>,
applicability: Applicability
) -> &mut Self

See Diagnostic::multipart_suggestion_verbose().

pub fn tool_only_multipart_suggestion(
&mut self,
msg: impl Into<SubdiagnosticMessage>,
suggestion: Vec<(Span, String)>,
applicability: Applicability
) -> &mut Self

See Diagnostic::tool_only_multipart_suggestion().

pub fn span_suggestion(
&mut self,
sp: Span,
msg: impl Into<SubdiagnosticMessage>,
suggestion: impl ToString,
applicability: Applicability
) -> &mut Self

See Diagnostic::span_suggestion().

pub fn span_suggestions(
&mut self,
sp: Span,
msg: impl Into<SubdiagnosticMessage>,
suggestions: impl IntoIterator<Item = String>,
applicability: Applicability
) -> &mut Self

See Diagnostic::span_suggestions().

pub fn multipart_suggestions(
&mut self,
msg: impl Into<SubdiagnosticMessage>,
suggestions: impl IntoIterator<Item = Vec<(Span, String)>>,
applicability: Applicability
) -> &mut Self

See Diagnostic::multipart_suggestions().

pub fn span_suggestion_short(
&mut self,
sp: Span,
msg: impl Into<SubdiagnosticMessage>,
suggestion: impl ToString,
applicability: Applicability
) -> &mut Self

See Diagnostic::span_suggestion_short().

pub fn span_suggestion_verbose(
&mut self,
sp: Span,
msg: impl Into<SubdiagnosticMessage>,
suggestion: impl ToString,
applicability: Applicability
) -> &mut Self

See Diagnostic::span_suggestion_verbose().

pub fn span_suggestion_hidden(
&mut self,
sp: Span,
msg: impl Into<SubdiagnosticMessage>,
suggestion: impl ToString,
applicability: Applicability
) -> &mut Self

See Diagnostic::span_suggestion_hidden().

pub fn tool_only_span_suggestion(
&mut self,
sp: Span,
msg: impl Into<SubdiagnosticMessage>,
suggestion: impl ToString,
applicability: Applicability
) -> &mut Self

See Diagnostic::tool_only_span_suggestion().

pub fn set_primary_message(
&mut self,
msg: impl Into<DiagnosticMessage>
) -> &mut Self

See Diagnostic::set_primary_message().

pub fn set_span(&mut self, sp: impl Into<MultiSpan>) -> &mut Self

See Diagnostic::set_span().

pub fn code(&mut self, s: DiagnosticId) -> &mut Self

See Diagnostic::code().

pub fn set_arg(
&mut self,
name: impl Into<Cow<'static, str>>,
arg: impl IntoDiagnosticArg
) -> &mut Self

See Diagnostic::set_arg().

pub fn subdiagnostic(&mut self, subdiagnostic: impl AddToDiagnostic) -> &mut Self

See Diagnostic::subdiagnostic().