Trait rustc_errors::CodeMapper

pub trait CodeMapper {
    fn lookup_char_pos(&self, pos: BytePos) -> Loc;

    fn span_to_lines(&self, sp: Span) -> FileLinesResult;

    fn span_to_string(&self, sp: Span) -> String;

    fn span_to_filename(&self, sp: Span) -> FileName;

    fn merge_spans(&self, sp_lhs: Span, sp_rhs: Span) -> Option<Span>;

    fn call_span_if_macro(&self, sp: Span) -> Span;

    fn ensure_filemap_source_present(&self, file_map: Rc<FileMap>) -> bool;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn lookup_char_pos(&self, pos: BytePos) -> Loc

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn span_to_lines(&self, sp: Span) -> FileLinesResult

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn span_to_string(&self, sp: Span) -> String

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn span_to_filename(&self, sp: Span) -> FileName

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn merge_spans(&self, sp_lhs: Span, sp_rhs: Span) -> Option<Span>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn call_span_if_macro(&self, sp: Span) -> Span

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn ensure_filemap_source_present(&self, file_map: Rc<FileMap>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?