rustc_errors::translation
Function to_fluent_args

pub fn to_fluent_args<'iter>(
    iter: impl Iterator<Item = DiagArg<'iter>>,
) -> FluentArgs<'static>

Convert diagnostic arguments (a rustc internal type that exists to implement Encodable/Decodable) into FluentArgs which is necessary to perform translation.

Typically performed once for each diagnostic at the start of emit_diagnostic and then passed around as a reference thereafter.