Crate rustc_errors

Diagnostics creation and emission for rustc.

This module contains the code for creating and emitting diagnostics.

Re-exports

pub use emitter::ColorConfig;

pub use rustc_error_messages::fluent_bundle;

Modules

annotate_snippet_emitter_writer

Emit diagnostics using the annotate-snippets library

diagnostic 🔒

diagnostic_builder 🔒

diagnostic_impls 🔒

emitter

The current rustc diagnostics emitter.

error

json

A JSON emitter for errors.

lock 🔒

Bindings to acquire a global named lock.

markdown

A simple markdown parser that can write formatted text to the terminal

registry

snippet 🔒

styled_buffer 🔒

translation

Macros

error_code

pluralize

struct_span_err

Structs

CodeSuggestion

DelayDm

A workaround for “good path” ICEs when formatting types in disabled lints.

DelayedBugPanic

Signifies that the compiler died with an explicit call to .delay_*_bug rather than a failed assertion, etc.

DelayedDiagnostic 🔒

Diagnostic

DiagnosticArgFromDisplay

DiagnosticBuilder

Used for emitting structured error messages and other diagnostic information.

DiagnosticStyledString

DiagnosticSymbolList

ErrorGuaranteed

Useful type to use with Result<> indicate that an error has already been reported to the user, so no need to continue checking.

ExpectedLifetimeParameter

ExplicitBug

Signifies that the compiler died with an explicit call to .bug or .span_bug rather than a failed assertion, etc.

FatalError

Used as a return value to signify a fatal error occurred. (It is also used as the argument to panic at the moment, but that will eventually not be true.)

FatalErrorMarker

Handler

A handler deals with errors and other compiler output. Certain errors (fatal, bug, unimpl) may cause immediate exit, others log errors for later reporting.

HandlerFlags

HandlerInner 🔒

This inner struct exists to keep it all behind a single lock; this is done to prevent possible deadlocks in a multi-threaded compiler, as well as inconsistent state observation.

IndicateAnonymousLifetime

InvalidFlushedDelayedDiagnosticLevel

LanguageIdentifier

LanguageIdentifier is a core struct representing a Unicode Language Identifier.

MultiSpan

A collection of Spans.

Noted

Marker type which enables implementation of create_note and emit_note functions for note-without-error struct diagnostics.

SingleLabelManySpans

Utility struct used to apply a single label while highlighting multiple spans

SpanLabel

A span together with some additional data.

SubDiagnostic

A “sub”-diagnostic attached to a parent diagnostic. For example, a note attached to an error.

Substitution

See the docs on CodeSuggestion::substitutions

SubstitutionHighlight

Used to translate between Spans and byte positions within a single output line in highlighted code of structured suggestions.

SubstitutionPart

Enums

Applicability

Indicates the confidence in the correctness of a suggestion.

DiagnosticArgValue

Simplified version of FluentValue that can implement Encodable and Decodable. Converted to a FluentValue by the emitter to be used in diagnostic translation.

DiagnosticId

DiagnosticMessage

Abstraction over a message in a diagnostic to support both translatable and non-translatable diagnostic messages.

LabelKind

The kind of label to attach when using SingleLabelManySpans

Level

StashKey

A key denoting where from a diagnostic was stashed.

Style

SubdiagnosticMessage

Abstraction over a message in a subdiagnostic (i.e. label, note, help, etc) to support both translatable and non-translatable diagnostic messages.

SuggestionStyle

TerminalUrl

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

TRACK_DIAGNOSTICS

Traits

AddToDiagnostic

Trait implemented by error types. This should not be implemented manually. Instead, use #[derive(Subdiagnostic)] – see rustc_macros::Subdiagnostic.

DecorateLint

Trait implemented by lint types. This should not be implemented manually. Instead, use #[derive(LintDiagnostic)] – see rustc_macros::LintDiagnostic.

EmissionGuarantee

Trait for types that DiagnosticBuilder::emit can return as a “guarantee” (or “proof”) token that the emission happened.

IntoDiagnostic

Trait implemented by error types. This should not be implemented manually. Instead, use #[derive(Diagnostic)] – see rustc_macros::Diagnostic.

IntoDiagnosticArg

Converts a value of a type into a DiagnosticArg (typically a field of an IntoDiagnostic struct). Implemented as a custom trait rather than From so that it is implemented on the type being converted rather than on DiagnosticArgValue, which enables types from other rustc_* crates to implement this.

Functions

add_elided_lifetime_in_path_suggestion

default_track_diagnostic 🔒

fallback_fluent_bundle

Return the default FluentBundle with standard “en-US” diagnostic messages.

fluent_bundle

Returns Fluent bundle with the user’s locale resources from $sysroot/share/locale/$requested_locale/*.ftl.

report_ambiguity_error

Type Aliases

DiagnosticArg

Simplified version of FluentArg that can implement Encodable and Decodable. Collection of DiagnosticArg are converted to FluentArgs (consuming the collection) at the start of diagnostic emission.

FluentBundle

LazyFallbackBundle

Type alias for the result of fallback_fluent_bundle - a reference-counted pointer to a lazily evaluated fluent bundle.

PErr

PResult