Crate rustc_errors

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Reexports

pub use emitter::ColorConfig;

Modules

emitter 	 [
Experimental
]

registry 	 [
Experimental
]

Structs

CodeSuggestion 	 [
Experimental
]

Diagnostic 	 [
Experimental
]

DiagnosticBuilder 	 [
Experimental
]
Used for emitting structured error messages and other diagnostic information.

DiagnosticStyledString 	 [
Experimental
]

ExplicitBug 	 [
Experimental
]
Signifies that the compiler died with an explicit call to .bug or .span_bug rather than a failed assertion, etc.

FatalError 	 [
Experimental
]
Used as a return value to signify a fatal error occurred. (It is also used as the argument to panic at the moment, but that will eventually not be true.)

Handler 	 [
Experimental
]
A handler deals with errors; certain errors (fatal, bug, unimpl) may cause immediate exit, others log errors for later reporting.

SubDiagnostic 	 [
Experimental
]
For example a note attached to an error.

Substitution 	 [
Experimental
]
See the docs on CodeSuggestion::substitutions

Enums

Level 	 [
Experimental
]

RenderSpan 	 [
Experimental
]

Traits

CodeMapper 	 [
Experimental
]