Crate rustc_errors

Re-exports

pub use emitter::ColorConfig;

Modules

emitter

registry

Structs

CodeSuggestion

Diagnostic

DiagnosticBuilder
Used for emitting structured error messages and other diagnostic information.

DiagnosticStyledString

ExplicitBug
Signifies that the compiler died with an explicit call to .bug or .span_bug rather than a failed assertion, etc.

FatalError
Used as a return value to signify a fatal error occurred. (It is also used as the argument to panic at the moment, but that will eventually not be true.)

FatalErrorMarker

Handler
A handler deals with errors; certain errors (fatal, bug, unimpl) may cause immediate exit, others log errors for later reporting.

HandlerFlags

SubDiagnostic
For example a note attached to an error.

Substitution
See the docs on CodeSuggestion::substitutions

SubstitutionPart

Enums

Applicability
Indicates the confidence in the correctness of a suggestion.

DiagnosticId

Level

SuggestionStyle

Constants

TRACK_DIAGNOSTICS

Traits

SourceMapper

Type Definitions

SourceMapperDyn