Module rustc_errors::emitter

The current rustc diagnostics emitter.

An Emitter takes care of generating the output from a DiagnosticBuilder struct.

There are various Emitter implementations that generate different output formats such as JSON and human readable output.

The output types are defined in rustc_session::config::ErrorOutputType.

Structs

Buffy ğŸ”’

EmitterWriter

Handles the writing of HumanReadableErrorType::Default and HumanReadableErrorType::Short

FileWithAnnotatedLines

Margin ğŸ”’

SilentEmitter

An emitter that does nothing when emitting a non-fatal diagnostic. Fatal diagnostics are forwarded to fatal_handler to avoid silent failures of rustc, as witnessed e.g. in issue #89358.

Enums

ColorConfig

DisplaySuggestion ğŸ”’

HumanReadableErrorType

Describes the way the content of the rendered field of the json output is generated

Constants

ANONYMIZED_LINE_NUM ğŸ”’

DEFAULT_COLUMN_WIDTH ğŸ”’

Default column width, used in tests and when terminal dimensions cannot be determined.

MAX_SUGGESTIONS

Maximum number of suggestions to be shown

OUTPUT_REPLACEMENTS ğŸ”’

Traits

Emitter

Emitter trait for emitting errors.

Functions

draw_col_separator ğŸ”’

draw_col_separator_no_space ğŸ”’

draw_col_separator_no_space_with_style ğŸ”’

draw_multiline_line ğŸ”’

draw_note_separator ğŸ”’

draw_range ğŸ”’

emit_to_destination ğŸ”’

from_stderr ğŸ”’

is_case_difference

Whether the original and suggested code are visually similar enough to warrant extra wording.

normalize_whitespace ğŸ”’

num_decimal_digits ğŸ”’

num_overlap ğŸ”’

overlaps ğŸ”’

Type Aliases

Destination