Module rustc_errors::emitter

Unstable (rustc_private)

Structs

BufferedWriter 	 [Unstable]
Buffered writer gives us a way on Unix to buffer up an entire error message before we output it. This helps to prevent interleaving of multiple error messages when multiple compiler processes error simultaneously

EmitterWriter 	 [Unstable]

Enums

ColorConfig 	 [Unstable]

Destination 	 [Unstable]

Constants

MAX_HIGHLIGHT_LINES 	 [Unstable]
maximum number of lines we will print for each error; arbitrary.

Traits

Emitter 	 [Unstable]
Emitter trait for emitting errors.

Type Definitions

BufferedStderr 	 [Unstable]