rustc_errors

Module emitter

The current rustc diagnostics emitter.

An Emitter takes care of generating the output from a Diag struct.

There are various Emitter implementations that generate different output formats such as JSON and human readable output.

The output types are defined in rustc_session::config::ErrorOutputType.

Structs

Buffy 🔒

FatalOnlyEmitter

An emitter that does nothing when emitting a non-fatal diagnostic. Fatal diagnostics are forwarded to fatal_emitter to avoid silent failures of rustc, as witnessed e.g. in issue #89358.

FileWithAnnotatedLines 🔒

HumanEmitter

Handles the writing of HumanReadableErrorType::Default and HumanReadableErrorType::Short

Margin 🔒

SilentEmitter

UnderlineParts 🔒

Enums

ColorConfig

ConfusionType

Represents the type of confusion detected between original and suggested code.

DisplaySuggestion 🔒

HumanReadableErrorType

Describes the way the content of the rendered field of the json output is generated

OutputTheme

TimingEvent

Constants

ANONYMIZED_LINE_NUM 🔒

BRIGHT_BLUE 🔒

On Windows, BRIGHT_BLUE is hard to read on black. Use cyan instead.

DEFAULT_COLUMN_WIDTH 🔒

Default column width, used in tests and when terminal dimensions cannot be determined.

MAX_SUGGESTIONS

Maximum number of suggestions to be shown

OUTPUT_REPLACEMENTS 🔒

Traits

Emitter

Emitter trait for emitting errors and other structured information.

Functions

detect_confusion_type

Whether the original and suggested code are visually similar enough to warrant extra wording.

emit_to_destination 🔒

is_different

Whether the original and suggested code are the same.

normalize_whitespace 🔒

num_decimal_digits 🔒

num_overlap 🔒

overlaps 🔒

should_show_source_code 🔒

stderr_destination

Type Aliases

Destination

DynEmitter