Struct rustc_errors::emitter::EmitterWriter

pub struct EmitterWriter {
    dst: Destination,
    sm: Option<Lrc<SourceMap>>,
    fluent_bundle: Option<Lrc<FluentBundle>>,
    fallback_bundle: LazyFallbackBundle,
    short_message: bool,
    teach: bool,
    ui_testing: bool,
    diagnostic_width: Option<usize>,
    macro_backtrace: bool,
    track_diagnostics: bool,
    terminal_url: TerminalUrl,
}

Handles the writing of HumanReadableErrorType::Default and HumanReadableErrorType::Short

Fields
dst: Destinationsm: Option<Lrc<SourceMap>>fluent_bundle: Option<Lrc<FluentBundle>>fallback_bundle: LazyFallbackBundleshort_message: boolteach: boolui_testing: booldiagnostic_width: Option<usize>macro_backtrace: booltrack_diagnostics: boolterminal_url: TerminalUrl
Implementations

impl EmitterWriter

pub fn sm(self, value: Option<Lrc<SourceMap>>) -> Self

pub fn fluent_bundle(self, value: Option<Lrc<FluentBundle>>) -> Self

pub fn short_message(self, value: bool) -> Self

pub fn teach(self, value: bool) -> Self

pub fn ui_testing(self, value: bool) -> Self

pub fn diagnostic_width(self, value: Option<usize>) -> Self

pub fn macro_backtrace(self, value: bool) -> Self

pub fn track_diagnostics(self, value: bool) -> Self

pub fn terminal_url(self, value: TerminalUrl) -> Self

impl EmitterWriter

pub fn stderr( color_config: ColorConfig, fallback_bundle: LazyFallbackBundle ) -> EmitterWriter

fn create( dst: Destination, fallback_bundle: LazyFallbackBundle ) -> EmitterWriter

pub fn new( dst: Box<dyn WriteColor + Send>, fallback_bundle: LazyFallbackBundle ) -> EmitterWriter

fn maybe_anonymized(&self, line_num: usize) -> Cow<'static, str>

fn draw_line( &self, buffer: &mut StyledBuffer, source_string: &str, line_index: usize, line_offset: usize, width_offset: usize, code_offset: usize, margin: Margin )

fn render_source_line( &self, buffer: &mut StyledBuffer, file: Lrc<SourceFile>, line: &Line, width_offset: usize, code_offset: usize, margin: Margin ) -> Vec<(usize, Style)>

fn get_multispan_max_line_num(&mut self, msp: &MultiSpan) -> usize

fn get_max_line_num( &mut self, span: &MultiSpan, children: &[SubDiagnostic] ) -> usize

fn msg_to_buffer( &self, buffer: &mut StyledBuffer, msg: &[(DiagnosticMessage, Style)], args: &FluentArgs<'_>, padding: usize, label: &str, override_style: Option<Style> )

Adds a left margin to every line but the first, given a padding length and the label being displayed, keeping the provided highlighting.

fn emit_message_default( &mut self, msp: &MultiSpan, msg: &[(DiagnosticMessage, Style)], args: &FluentArgs<'_>, code: &Option<DiagnosticId>, level: &Level, max_line_num_len: usize, is_secondary: bool, emitted_at: Option<&DiagnosticLocation> ) -> Result<()>

fn emit_suggestion_default( &mut self, span: &MultiSpan, suggestion: &CodeSuggestion, args: &FluentArgs<'_>, level: &Level, max_line_num_len: usize ) -> Result<()>

fn emit_messages_default( &mut self, level: &Level, message: &[(DiagnosticMessage, Style)], args: &FluentArgs<'_>, code: &Option<DiagnosticId>, span: &MultiSpan, children: &[SubDiagnostic], suggestions: &[CodeSuggestion], emitted_at: Option<&DiagnosticLocation> )

fn draw_code_line( &self, buffer: &mut StyledBuffer, row_num: &mut usize, highlight_parts: &[SubstitutionHighlight], line_num: usize, line_to_add: &str, show_code_change: DisplaySuggestion, max_line_num_len: usize, file_lines: &FileLines, is_multiline: bool )