Struct rustc_errors::emitter::EmitterWriter

pub struct EmitterWriter { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl EmitterWriter

pub fn stderr(
color_config: ColorConfig,
code_map: Option<Rc<CodeMapper>>
) -> EmitterWriter

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn new(
dst: Box<Write + Send>,
code_map: Option<Rc<CodeMapper>>
) -> EmitterWriter

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?