rustc_errors::emitter
Struct SilentEmitter

pub struct SilentEmitter {
    pub fallback_bundle: LazyFallbackBundle,
    pub fatal_dcx: DiagCtxt,
    pub fatal_note: Option<String>,
    pub emit_fatal_diagnostic: bool,
}

An emitter that does nothing when emitting a non-fatal diagnostic. Fatal diagnostics are forwarded to fatal_dcx to avoid silent failures of rustc, as witnessed e.g. in issue #89358.

Fields
fallback_bundle: LazyFallbackBundlefatal_dcx: DiagCtxtfatal_note: Option<String>emit_fatal_diagnostic: bool