Trait rustc_errors::emitter::Emitter

pub trait Emitter {
    fn emit(&mut self, db: &DiagnosticBuilder);
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Emitter trait for emitting errors.

Required Methods

fn emit(&mut self, db: &DiagnosticBuilder)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Emit a structured diagnostic.