Struct rustc_errors::Diagnostic

pub struct Diagnostic {
    pub(crate) level: Level,
    pub message: Vec<(DiagnosticMessage, Style)>,
    pub code: Option<DiagnosticId>,
    pub span: MultiSpan,
    pub children: Vec<SubDiagnostic>,
    pub suggestions: Result<Vec<CodeSuggestion>, SuggestionsDisabled>,
    args: FxHashMap<Cow<'static, str>, DiagnosticArgValue<'static>>,
    pub sort_span: Span,
    pub is_lint: bool,
    pub emitted_at: DiagnosticLocation,
}

Fields
level: Levelmessage: Vec<(DiagnosticMessage, Style)>code: Option<DiagnosticId>span: MultiSpanchildren: Vec<SubDiagnostic>suggestions: Result<Vec<CodeSuggestion>, SuggestionsDisabled>args: FxHashMap<Cow<'static, str>, DiagnosticArgValue<'static>>sort_span: Span

This is not used for highlighting or rendering any error message. Rather, it can be used as a sort key to sort a buffer of diagnostics. By default, it is the primary span of span if there is one. Otherwise, it is DUMMY_SP.

is_lint: bool

If diagnostic is from Lint, custom hash function ignores notes otherwise hash is based on the all the fields

emitted_at: DiagnosticLocation

With -Ztrack_diagnostics enabled, we print where in rustc this error was emitted.

Implementations

impl Diagnostic

pub fn new<M: Into<DiagnosticMessage>>(level: Level, message: M) -> Self

pub fn new_with_messages( level: Level, messages: Vec<(DiagnosticMessage, Style)> ) -> Self

pub fn new_with_code<M: Into<DiagnosticMessage>>( level: Level, code: Option<DiagnosticId>, message: M ) -> Self

pub fn level(&self) -> Level

pub fn is_error(&self) -> bool

pub fn update_unstable_expectation_id( &mut self, unstable_to_stable: &FxHashMap<LintExpectationId, LintExpectationId> )

pub fn has_future_breakage(&self) -> bool

pub fn is_force_warn(&self) -> bool

pub fn downgrade_to_delayed_bug(&mut self) -> &mut Self

Delay emission of this diagnostic as a bug.

This can be useful in contexts where an error indicates a bug but typically this only happens when other compilation errors have already happened. In those cases this can be used to defer emission of this diagnostic as a bug in the compiler only if no other errors have been emitted.

In the meantime, though, callsites are required to deal with the “bug” locally in whichever way makes the most sense.

pub fn span_label( &mut self, span: Span, label: impl Into<SubdiagnosticMessage> ) -> &mut Self

Adds a span/label to be included in the resulting snippet.

This is pushed onto the MultiSpan that was created when the diagnostic was first built. That means it will be shown together with the original span/label, not a span added by one of the span_{note,warn,help,suggestions} methods.

This span is not considered a “primary span”; only the Span supplied when creating the diagnostic is primary.

pub fn span_labels( &mut self, spans: impl IntoIterator<Item = Span>, label: &str ) -> &mut Self

Labels all the given spans with the provided label. See Self::span_label() for more information.

pub fn replace_span_with(&mut self, after: Span, keep_label: bool) -> &mut Self

pub fn note_expected_found( &mut self, expected_label: &dyn Display, expected: DiagnosticStyledString, found_label: &dyn Display, found: DiagnosticStyledString ) -> &mut Self

pub fn note_unsuccessful_coercion( &mut self, expected: DiagnosticStyledString, found: DiagnosticStyledString ) -> &mut Self

pub fn note_expected_found_extra( &mut self, expected_label: &dyn Display, expected: DiagnosticStyledString, found_label: &dyn Display, found: DiagnosticStyledString, expected_extra: &dyn Display, found_extra: &dyn Display ) -> &mut Self

pub fn note_trait_signature( &mut self, name: Symbol, signature: String ) -> &mut Self

pub fn note(&mut self, msg: impl Into<SubdiagnosticMessage>) -> &mut Self

Add a note attached to this diagnostic.

pub fn highlighted_note<M: Into<SubdiagnosticMessage>>( &mut self, msg: Vec<(M, Style)> ) -> &mut Self

pub fn note_once(&mut self, msg: impl Into<SubdiagnosticMessage>) -> &mut Self

Prints the span with a note above it. This is like Diagnostic::note(), but it gets its own span.

pub fn span_note<S: Into<MultiSpan>>( &mut self, sp: S, msg: impl Into<SubdiagnosticMessage> ) -> &mut Self

Prints the span with a note above it. This is like Diagnostic::note(), but it gets its own span.

pub fn span_note_once<S: Into<MultiSpan>>( &mut self, sp: S, msg: impl Into<SubdiagnosticMessage> ) -> &mut Self

Prints the span with a note above it. This is like Diagnostic::note(), but it gets its own span.

pub fn warn(&mut self, msg: impl Into<SubdiagnosticMessage>) -> &mut Self

Add a warning attached to this diagnostic.

pub fn span_warn<S: Into<MultiSpan>>( &mut self, sp: S, msg: impl Into<SubdiagnosticMessage> ) -> &mut Self

Prints the span with a warning above it. This is like Diagnostic::warn(), but it gets its own span.

pub fn help(&mut self, msg: impl Into<SubdiagnosticMessage>) -> &mut Self

Add a help message attached to this diagnostic.

pub fn highlighted_help(&mut self, msg: Vec<(String, Style)>) -> &mut Self

Add a help message attached to this diagnostic with a customizable highlighted message.

pub fn span_help<S: Into<MultiSpan>>( &mut self, sp: S, msg: impl Into<SubdiagnosticMessage> ) -> &mut Self

Prints the span with some help above it. This is like Diagnostic::help(), but it gets its own span.

pub fn disable_suggestions(&mut self) -> &mut Self

Disallow attaching suggestions this diagnostic. Any suggestions attached e.g. with the span_suggestion_* methods (before and after the call to disable_suggestions) will be ignored.

pub fn clear_suggestions(&mut self) -> &mut Self

Clear any existing suggestions.

fn push_suggestion(&mut self, suggestion: CodeSuggestion)

Helper for pushing to self.suggestions, if available (not disable).

pub fn multipart_suggestion( &mut self, msg: impl Into<SubdiagnosticMessage>, suggestion: Vec<(Span, String)>, applicability: Applicability ) -> &mut Self

Show a suggestion that has multiple parts to it. In other words, multiple changes need to be applied as part of this suggestion.

pub fn multipart_suggestion_verbose( &mut self, msg: impl Into<SubdiagnosticMessage>, suggestion: Vec<(Span, String)>, applicability: Applicability ) -> &mut Self

Show a suggestion that has multiple parts to it, always as it’s own subdiagnostic. In other words, multiple changes need to be applied as part of this suggestion.

pub fn multipart_suggestion_with_style( &mut self, msg: impl Into<SubdiagnosticMessage>, suggestion: Vec<(Span, String)>, applicability: Applicability, style: SuggestionStyle ) -> &mut Self

Diagnostic::multipart_suggestion() but you can set the SuggestionStyle.

pub fn tool_only_multipart_suggestion( &mut self, msg: impl Into<SubdiagnosticMessage>, suggestion: Vec<(Span, String)>, applicability: Applicability ) -> &mut Self

Prints out a message with for a multipart suggestion without showing the suggested code.

This is intended to be used for suggestions that are obvious in what the changes need to be from the message, showing the span label inline would be visually unpleasant (marginally overlapping spans or multiline spans) and showing the snippet window wouldn’t improve understandability.

pub fn span_suggestion( &mut self, sp: Span, msg: impl Into<SubdiagnosticMessage>, suggestion: impl ToString, applicability: Applicability ) -> &mut Self

Prints out a message with a suggested edit of the code.

In case of short messages and a simple suggestion, rustc displays it as a label:

try adding parentheses: `(tup.0).1`

The message

should not end in any punctuation (a : is added automatically)

should not be a question (avoid language like “did you mean”)

should not contain any phrases like “the following”, “as shown”, etc.

may look like “to do xyz, use” or “to do xyz, use abc”

may contain a name of a function, variable, or type, but not whole expressions

See CodeSuggestion for more information.

pub fn span_suggestion_with_style( &mut self, sp: Span, msg: impl Into<SubdiagnosticMessage>, suggestion: impl ToString, applicability: Applicability, style: SuggestionStyle ) -> &mut Self

Diagnostic::span_suggestion() but you can set the SuggestionStyle.

pub fn span_suggestion_verbose( &mut self, sp: Span, msg: impl Into<SubdiagnosticMessage>, suggestion: impl ToString, applicability: Applicability ) -> &mut Self

Always show the suggested change.

pub fn span_suggestions( &mut self, sp: Span, msg: impl Into<SubdiagnosticMessage>, suggestions: impl IntoIterator<Item = String>, applicability: Applicability ) -> &mut Self

Prints out a message with multiple suggested edits of the code. See also Diagnostic::span_suggestion().

pub fn span_suggestions_with_style( &mut self, sp: Span, msg: impl Into<SubdiagnosticMessage>, suggestions: impl IntoIterator<Item = String>, applicability: Applicability, style: SuggestionStyle ) -> &mut Self

Diagnostic::span_suggestions() but you can set the SuggestionStyle.

pub fn multipart_suggestions( &mut self, msg: impl Into<SubdiagnosticMessage>, suggestions: impl IntoIterator<Item = Vec<(Span, String)>>, applicability: Applicability ) -> &mut Self

Prints out a message with multiple suggested edits of the code, where each edit consists of multiple parts. See also Diagnostic::multipart_suggestion().

pub fn span_suggestion_short( &mut self, sp: Span, msg: impl Into<SubdiagnosticMessage>, suggestion: impl ToString, applicability: Applicability ) -> &mut Self

Prints out a message with a suggested edit of the code. If the suggestion is presented inline, it will only show the message and not the suggestion.

See CodeSuggestion for more information.

pub fn span_suggestion_hidden( &mut self, sp: Span, msg: impl Into<SubdiagnosticMessage>, suggestion: impl ToString, applicability: Applicability ) -> &mut Self

Prints out a message for a suggestion without showing the suggested code.

This is intended to be used for suggestions that are obvious in what the changes need to be from the message, showing the span label inline would be visually unpleasant (marginally overlapping spans or multiline spans) and showing the snippet window wouldn’t improve understandability.

pub fn tool_only_span_suggestion( &mut self, sp: Span, msg: impl Into<SubdiagnosticMessage>, suggestion: impl ToString, applicability: Applicability ) -> &mut Self

Adds a suggestion to the JSON output that will not be shown in the CLI.

This is intended to be used for suggestions that are very obvious in what the changes need to be from the message, but we still want other tools to be able to apply them.

pub fn subdiagnostic( &mut self, subdiagnostic: impl AddToDiagnostic ) -> &mut Self

Add a subdiagnostic from a type that implements Subdiagnostic (see rustc_macros::Subdiagnostic).

pub fn eager_subdiagnostic( &mut self, handler: &Handler, subdiagnostic: impl AddToDiagnostic ) -> &mut Self

Add a subdiagnostic from a type that implements Subdiagnostic (see rustc_macros::Subdiagnostic). Performs eager translation of any translatable messages used in the subdiagnostic, so suitable for use with repeated messages (i.e. re-use of interpolated variables).

pub fn set_span<S: Into<MultiSpan>>(&mut self, sp: S) -> &mut Self

pub fn set_is_lint(&mut self) -> &mut Self

pub fn code(&mut self, s: DiagnosticId) -> &mut Self

pub fn clear_code(&mut self) -> &mut Self

pub fn get_code(&self) -> Option<DiagnosticId>

pub fn set_primary_message( &mut self, msg: impl Into<DiagnosticMessage> ) -> &mut Self

pub fn args(&self) -> impl Iterator<Item = DiagnosticArg<'_, 'static>>

pub fn set_arg( &mut self, name: impl Into<Cow<'static, str>>, arg: impl IntoDiagnosticArg ) -> &mut Self

pub fn replace_args( &mut self, args: FxHashMap<Cow<'static, str>, DiagnosticArgValue<'static>> )

pub fn styled_message(&self) -> &[(DiagnosticMessage, Style)]

pub(crate) fn subdiagnostic_message_to_diagnostic_message( &self, attr: impl Into<SubdiagnosticMessage> ) -> DiagnosticMessage

Helper function that takes a SubdiagnosticMessage and returns a DiagnosticMessage by combining it with the primary message of the diagnostic (if translatable, otherwise it just passes the user’s string along).

pub fn sub( &mut self, level: Level, message: impl Into<SubdiagnosticMessage>, span: MultiSpan, render_span: Option<MultiSpan> )

Convenience function for internal use, clients should use one of the public methods above.

Used by proc_macro_server for implementing server::Diagnostic.

fn sub_with_highlights<M: Into<SubdiagnosticMessage>>( &mut self, level: Level, message: Vec<(M, Style)>, span: MultiSpan, render_span: Option<MultiSpan> )

Convenience function for internal use, clients should use one of the public methods above.

fn keys( &self ) -> (&Level, &[(DiagnosticMessage, Style)], Vec<(&Cow<'static, str>, &DiagnosticArgValue<'static>)>, &Option<DiagnosticId>, &MultiSpan, &Result<Vec<CodeSuggestion>, SuggestionsDisabled>, Option<&[SubDiagnostic]>)

Fields used for Hash, and PartialEq trait