Struct rustc_errors::Handler

pub struct Handler {
    pub(crate) inner: Lock<HandlerInner>,
}

A handler deals with errors and other compiler output. Certain errors (fatal, bug, unimpl) may cause immediate exit, others log errors for later reporting.

Fields
inner: Lock<HandlerInner>
Implementations

impl Handler

pub fn with_tty_emitter( sm: Option<Lrc<SourceMap>>, fallback_bundle: LazyFallbackBundle ) -> Self

pub fn disable_warnings(self) -> Self

pub fn treat_err_as_bug(self, treat_err_as_bug: NonZeroUsize) -> Self

pub fn with_flags(self, flags: HandlerFlags) -> Self

pub fn with_ice_file(self, ice_file: PathBuf) -> Self

pub fn with_emitter(emitter: Box<dyn Emitter + Send>) -> Self

pub fn eagerly_translate<'a>( &self, message: DiagnosticMessage, args: impl Iterator<Item = DiagnosticArg<'a, 'static>> ) -> SubdiagnosticMessage

Translate message eagerly with args to SubdiagnosticMessage::Eager.

pub fn eagerly_translate_to_string<'a>( &self, message: DiagnosticMessage, args: impl Iterator<Item = DiagnosticArg<'a, 'static>> ) -> String

Translate message eagerly with args to String.

pub fn can_emit_warnings(&self) -> bool

pub fn reset_err_count(&self)

Resets the diagnostic error count as well as the cached emitted diagnostics.

NOTE: do not call this function from rustc. It is only meant to be called from external tools that want to reuse a Parser cleaning the previously emitted diagnostics as well as the overall count of emitted error diagnostics.

pub fn stash_diagnostic(&self, span: Span, key: StashKey, diag: Diagnostic)

Stash a given diagnostic with the given Span and StashKey as the key. Retrieve a stashed diagnostic with steal_diagnostic.

pub fn steal_diagnostic( &self, span: Span, key: StashKey ) -> Option<DiagnosticBuilder<'_, ()>>

Steal a previously stashed diagnostic with the given Span and StashKey as the key.

pub fn has_stashed_diagnostic(&self, span: Span, key: StashKey) -> bool

pub fn emit_stashed_diagnostics(&self) -> Option<ErrorGuaranteed>

Emit all stashed diagnostics.

pub fn struct_diagnostic<G: EmissionGuarantee>( &self, msg: impl Into<DiagnosticMessage> ) -> DiagnosticBuilder<'_, G>

Construct a builder with the msg at the level appropriate for the specific EmissionGuarantee.

pub fn struct_span_warn( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage> ) -> DiagnosticBuilder<'_, ()>

Construct a builder at the Warning level at the given span and with the msg.

Attempting to .emit() the builder will only emit if either:

can_emit_warnings is true

is_force_warn was set in DiagnosticId::Lint

pub fn struct_span_warn_with_expectation( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage>, id: LintExpectationId ) -> DiagnosticBuilder<'_, ()>

Construct a builder at the Warning level at the given span and with the msg. The id is used for lint emissions which should also fulfill a lint expectation.

Attempting to .emit() the builder will only emit if either:

can_emit_warnings is true

is_force_warn was set in DiagnosticId::Lint

pub fn struct_span_allow( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage> ) -> DiagnosticBuilder<'_, ()>

Construct a builder at the Allow level at the given span and with the msg.

pub fn struct_span_warn_with_code( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage>, code: DiagnosticId ) -> DiagnosticBuilder<'_, ()>

Construct a builder at the Warning level at the given span and with the msg. Also include a code.

pub fn struct_warn( &self, msg: impl Into<DiagnosticMessage> ) -> DiagnosticBuilder<'_, ()>

Construct a builder at the Warning level with the msg.

Attempting to .emit() the builder will only emit if either:

can_emit_warnings is true

is_force_warn was set in DiagnosticId::Lint

pub fn struct_warn_with_expectation( &self, msg: impl Into<DiagnosticMessage>, id: LintExpectationId ) -> DiagnosticBuilder<'_, ()>

Construct a builder at the Warning level with the msg. The id is used for lint emissions which should also fulfill a lint expectation.

Attempting to .emit() the builder will only emit if either:

can_emit_warnings is true

is_force_warn was set in DiagnosticId::Lint

pub fn struct_allow( &self, msg: impl Into<DiagnosticMessage> ) -> DiagnosticBuilder<'_, ()>

Construct a builder at the Allow level with the msg.

pub fn struct_expect( &self, msg: impl Into<DiagnosticMessage>, id: LintExpectationId ) -> DiagnosticBuilder<'_, ()>

Construct a builder at the Expect level with the msg.

pub fn struct_span_err( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage> ) -> DiagnosticBuilder<'_, ErrorGuaranteed>

Construct a builder at the Error level at the given span and with the msg.

pub fn struct_span_err_with_code( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage>, code: DiagnosticId ) -> DiagnosticBuilder<'_, ErrorGuaranteed>

Construct a builder at the Error level at the given span, with the msg, and code.

pub fn struct_err( &self, msg: impl Into<DiagnosticMessage> ) -> DiagnosticBuilder<'_, ErrorGuaranteed>

Construct a builder at the Error level with the msg.

pub fn struct_err_with_code( &self, msg: impl Into<DiagnosticMessage>, code: DiagnosticId ) -> DiagnosticBuilder<'_, ErrorGuaranteed>

Construct a builder at the Error level with the msg and the code.

pub fn struct_warn_with_code( &self, msg: impl Into<DiagnosticMessage>, code: DiagnosticId ) -> DiagnosticBuilder<'_, ()>

Construct a builder at the Warn level with the msg and the code.

pub fn struct_span_fatal( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage> ) -> DiagnosticBuilder<'_, !>

Construct a builder at the Fatal level at the given span and with the msg.

pub fn struct_span_fatal_with_code( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage>, code: DiagnosticId ) -> DiagnosticBuilder<'_, !>

Construct a builder at the Fatal level at the given span, with the msg, and code.

pub fn struct_fatal( &self, msg: impl Into<DiagnosticMessage> ) -> DiagnosticBuilder<'_, !>

Construct a builder at the Error level with the msg.

pub fn struct_help( &self, msg: impl Into<DiagnosticMessage> ) -> DiagnosticBuilder<'_, ()>

Construct a builder at the Help level with the msg.

pub fn struct_note_without_error( &self, msg: impl Into<DiagnosticMessage> ) -> DiagnosticBuilder<'_, ()>

Construct a builder at the Note level with the msg.

pub fn span_fatal( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage> ) -> !

pub fn span_fatal_with_code( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage>, code: DiagnosticId ) -> !

pub fn span_err( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage> ) -> ErrorGuaranteed

pub fn span_err_with_code( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage>, code: DiagnosticId )

pub fn span_warn( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage> )

pub fn span_warn_with_code( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage>, code: DiagnosticId )

pub fn span_bug(&self, span: impl Into<MultiSpan>, msg: impl Into<String>) -> !

pub fn delay_span_bug( &self, span: impl Into<MultiSpan>, msg: impl Into<String> ) -> ErrorGuaranteed

For documentation on this, see Session::delay_span_bug.

pub fn delay_good_path_bug(&self, msg: impl Into<DiagnosticMessage>)

pub fn span_bug_no_panic( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage> )

pub fn span_note_without_error( &self, span: impl Into<MultiSpan>, msg: impl Into<DiagnosticMessage> )

pub fn span_note_diag( &self, span: Span, msg: impl Into<DiagnosticMessage> ) -> DiagnosticBuilder<'_, ()>

pub fn fatal(&self, msg: impl Into<DiagnosticMessage>) -> FatalError

pub fn err(&self, msg: impl Into<DiagnosticMessage>) -> ErrorGuaranteed

pub fn warn(&self, msg: impl Into<DiagnosticMessage>)

pub fn note_without_error(&self, msg: impl Into<DiagnosticMessage>)

pub fn bug(&self, msg: impl Into<DiagnosticMessage>) -> !

pub fn err_count(&self) -> usize

pub fn has_errors(&self) -> Option<ErrorGuaranteed>

pub fn has_errors_or_lint_errors(&self) -> Option<ErrorGuaranteed>

pub fn has_errors_or_delayed_span_bugs(&self) -> Option<ErrorGuaranteed>

pub fn is_compilation_going_to_fail(&self) -> Option<ErrorGuaranteed>

pub fn print_error_count(&self, registry: &Registry)

pub fn take_future_breakage_diagnostics(&self) -> Vec<Diagnostic>

pub fn abort_if_errors(&self)

pub fn must_teach(&self, code: &DiagnosticId) -> bool

true if we haven’t taught a diagnostic with this code already. The caller must then teach the user about such a diagnostic.

Used to suppress emitting the same error multiple times with extended explanation when calling -Zteach.

pub fn force_print_diagnostic(&self, db: Diagnostic)

pub fn emit_diagnostic( &self, diagnostic: &mut Diagnostic ) -> Option<ErrorGuaranteed>

pub fn emit_err<'a>(&'a self, err: impl IntoDiagnostic<'a>) -> ErrorGuaranteed

pub fn create_err<'a>( &'a self, err: impl IntoDiagnostic<'a> ) -> DiagnosticBuilder<'a, ErrorGuaranteed>

pub fn create_warning<'a>( &'a self, warning: impl IntoDiagnostic<'a, ()> ) -> DiagnosticBuilder<'a, ()>

pub fn emit_warning<'a>(&'a self, warning: impl IntoDiagnostic<'a, ()>)

pub fn create_almost_fatal<'a>( &'a self, fatal: impl IntoDiagnostic<'a, FatalError> ) -> DiagnosticBuilder<'a, FatalError>

pub fn emit_almost_fatal<'a>( &'a self, fatal: impl IntoDiagnostic<'a, FatalError> ) -> FatalError

pub fn create_fatal<'a>( &'a self, fatal: impl IntoDiagnostic<'a, !> ) -> DiagnosticBuilder<'a, !>

pub fn emit_fatal<'a>(&'a self, fatal: impl IntoDiagnostic<'a, !>) -> !

pub fn create_bug<'a>( &'a self, bug: impl IntoDiagnostic<'a, Bug> ) -> DiagnosticBuilder<'a, Bug>

pub fn emit_bug<'a>(&'a self, bug: impl IntoDiagnostic<'a, Bug>) -> Bug

pub fn emit_note<'a>(&'a self, note: impl IntoDiagnostic<'a, Noted>) -> Noted

pub fn create_note<'a>( &'a self, note: impl IntoDiagnostic<'a, Noted> ) -> DiagnosticBuilder<'a, Noted>

pub(crate) fn emit_diag_at_span( &self, diag: Diagnostic, sp: impl Into<MultiSpan> ) -> Option<ErrorGuaranteed>

pub fn emit_artifact_notification(&self, path: &Path, artifact_type: &str)

pub fn emit_future_breakage_report(&self, diags: Vec<Diagnostic>)

pub fn emit_unused_externs( &self, lint_level: Level, loud: bool, unused_externs: &[&str] )

pub fn update_unstable_expectation_id( &self, unstable_to_stable: &FxHashMap<LintExpectationId, LintExpectationId> )

pub fn steal_fulfilled_expectation_ids(&self) -> FxHashSet<LintExpectationId>

This methods steals all LintExpectationIds that are stored inside HandlerInner and indicate that the linked expectation has been fulfilled.

pub fn flush_delayed(&self)