Struct rustc_errors::HandlerInner

pub(crate) struct HandlerInner {

    pub(crate) flags: HandlerFlags,
    pub(crate) lint_err_count: usize,
    pub(crate) err_count: usize,
    pub(crate) warn_count: usize,
    pub(crate) deduplicated_err_count: usize,
    pub(crate) emitter: IntoDynSyncSend<Box<dyn Emitter + Send>>,
    pub(crate) delayed_span_bugs: Vec<DelayedDiagnostic>,
    pub(crate) delayed_good_path_bugs: Vec<DelayedDiagnostic>,
    pub(crate) suppressed_expected_diag: bool,
    pub(crate) taught_diagnostics: FxHashSet<DiagnosticId>,
    pub(crate) emitted_diagnostic_codes: FxIndexSet<DiagnosticId>,
    pub(crate) emitted_diagnostics: FxHashSet<Hash128>,
    pub(crate) stashed_diagnostics: FxIndexMap<(Span, StashKey), Diagnostic>,
    pub(crate) deduplicated_warn_count: usize,
    pub(crate) future_breakage_diagnostics: Vec<Diagnostic>,
    pub(crate) check_unstable_expect_diagnostics: bool,
    pub(crate) unstable_expect_diagnostics: Vec<Diagnostic>,
    pub(crate) fulfilled_expectations: FxHashSet<LintExpectationId>,
    pub(crate) ice_file: Option<PathBuf>,

}

This inner struct exists to keep it all behind a single lock; this is done to prevent possible deadlocks in a multi-threaded compiler, as well as inconsistent state observation.

Fields
flags: HandlerFlagslint_err_count: usize

The number of lint errors that have been emitted.

err_count: usize

The number of errors that have been emitted, including duplicates.

This is not necessarily the count that’s reported to the user once compilation ends.

warn_count: usizededuplicated_err_count: usizeemitter: IntoDynSyncSend<Box<dyn Emitter + Send>>delayed_span_bugs: Vec<DelayedDiagnostic>delayed_good_path_bugs: Vec<DelayedDiagnostic>suppressed_expected_diag: bool

This flag indicates that an expected diagnostic was emitted and suppressed. This is used for the delayed_good_path_bugs check.

taught_diagnostics: FxHashSet<DiagnosticId>

This set contains the DiagnosticId of all emitted diagnostics to avoid emitting the same diagnostic with extended help (--teach) twice, which would be unnecessary repetition.

emitted_diagnostic_codes: FxIndexSet<DiagnosticId>

Used to suggest rustc –explain <error code>

emitted_diagnostics: FxHashSet<Hash128>

This set contains a hash of every diagnostic that has been emitted by this handler. These hashes is used to avoid emitting the same error twice.

stashed_diagnostics: FxIndexMap<(Span, StashKey), Diagnostic>

Stashed diagnostics emitted in one stage of the compiler that may be stolen by other stages (e.g. to improve them and add more information). The stashed diagnostics count towards the total error count. When .abort_if_errors() is called, these are also emitted.

deduplicated_warn_count: usize

The warning count, used for a recap upon finishing

future_breakage_diagnostics: Vec<Diagnostic>check_unstable_expect_diagnostics: bool

The Self::unstable_expect_diagnostics should be empty when this struct is dropped. However, it can have values if the compilation is stopped early or is only partially executed. To avoid ICEs, like in rust#94953 we only check if Self::unstable_expect_diagnostics is empty, if the expectation ids have been converted.

unstable_expect_diagnostics: Vec<Diagnostic>

Expected Diagnostics store a LintExpectationId as part of the lint level. LintExpectationIds created early during the compilation (before HirIds have been defined) are not stable and can therefore not be stored on disk. This buffer stores these diagnostics until the ID has been replaced by a stable LintExpectationId. The Diagnostics are the submitted for storage and added to the list of fulfilled expectations.

fulfilled_expectations: FxHashSet<LintExpectationId>

expected diagnostic will have the level Expect which additionally carries the LintExpectationId of the expectation that can be marked as fulfilled. This is a collection of all LintExpectationIds that have been marked as fulfilled this way.

ice_file: Option<PathBuf>

The file where the ICE information is stored. This allows delayed_span_bug backtraces to be stored along side the main panic backtrace.

Implementations

impl HandlerInner

pub(crate) fn must_teach(&mut self, code: &DiagnosticId) -> bool

pub(crate) fn force_print_diagnostic(&mut self, db: Diagnostic)

pub(crate) fn emit_stashed_diagnostics(&mut self) -> Option<ErrorGuaranteed>

Emit all stashed diagnostics.

pub(crate) fn emit_diagnostic( &mut self, diagnostic: &mut Diagnostic ) -> Option<ErrorGuaranteed>

pub(crate) fn emit_artifact_notification( &mut self, path: &Path, artifact_type: &str )

pub(crate) fn emit_unused_externs( &mut self, lint_level: Level, unused_externs: &[&str] )

pub(crate) fn treat_err_as_bug(&self) -> bool

pub(crate) fn delayed_bug_count(&self) -> usize

pub(crate) fn print_error_count(&mut self, registry: &Registry)

pub(crate) fn stash(&mut self, key: (Span, StashKey), diagnostic: Diagnostic)

pub(crate) fn steal(&mut self, key: (Span, StashKey)) -> Option<Diagnostic>

pub(crate) fn err_count(&self) -> usize

pub(crate) fn has_errors(&self) -> bool

pub(crate) fn has_errors_or_lint_errors(&self) -> bool

pub(crate) fn has_errors_or_delayed_span_bugs(&self) -> bool

pub(crate) fn has_any_message(&self) -> bool

pub(crate) fn is_compilation_going_to_fail(&self) -> bool

pub(crate) fn abort_if_errors(&mut self)

pub(crate) fn span_bug( &mut self, sp: impl Into<MultiSpan>, msg: impl Into<String> ) -> !

pub(crate) fn emit_diag_at_span( &mut self, diag: Diagnostic, sp: impl Into<MultiSpan> )

pub(crate) fn delay_span_bug( &mut self, sp: impl Into<MultiSpan>, msg: impl Into<String> ) -> ErrorGuaranteed

For documentation on this, see Session::delay_span_bug.

pub(crate) fn delay_good_path_bug(&mut self, msg: impl Into<DiagnosticMessage>)

pub(crate) fn failure(&mut self, msg: impl Into<DiagnosticMessage>)

pub(crate) fn fatal(&mut self, msg: impl Into<DiagnosticMessage>) -> FatalError

pub(crate) fn err( &mut self, msg: impl Into<DiagnosticMessage> ) -> ErrorGuaranteed

pub(crate) fn emit( &mut self, level: Level, msg: impl Into<DiagnosticMessage> ) -> ErrorGuaranteed

Emit an error; level should be Error or Fatal.

pub(crate) fn bug(&mut self, msg: impl Into<DiagnosticMessage>) -> !

pub(crate) fn flush_delayed( &mut self, bugs: impl IntoIterator<Item = DelayedDiagnostic>, explanation: impl Into<DiagnosticMessage> + Copy )

pub(crate) fn bump_lint_err_count(&mut self)

pub(crate) fn bump_err_count(&mut self)

pub(crate) fn bump_warn_count(&mut self)

pub(crate) fn panic_if_treat_err_as_bug(&self)