Struct rustc_errors::styled_buffer::StyledBuffer

pub struct StyledBuffer {
    lines: Vec<Vec<StyledChar>>,
}

Fields
lines: Vec<Vec<StyledChar>>
Implementations

impl StyledBuffer

pub fn new() -> StyledBuffer

pub fn render(&self) -> Vec<Vec<StyledString>>

Returns content of StyledBuffer split by lines and line styles

fn ensure_lines(&mut self, line: usize)

pub fn putc(&mut self, line: usize, col: usize, chr: char, style: Style)

Sets chr with style for given line, col. If line does not exist in our buffer, adds empty lines up to the given and fills the last line with unstyled whitespace.

pub fn puts(&mut self, line: usize, col: usize, string: &str, style: Style)

Sets string with style for given line, starting from col. If line does not exist in our buffer, adds empty lines up to the given and fills the last line with unstyled whitespace.

pub fn prepend(&mut self, line: usize, string: &str, style: Style)

For given line inserts string with style before old content of that line, adding lines if needed

pub fn append(&mut self, line: usize, string: &str, style: Style)

For given line inserts string with style after old content of that line, adding lines if needed

pub fn num_lines(&self) -> usize

pub fn set_style_range( &mut self, line: usize, col_start: usize, col_end: usize, style: Style, overwrite: bool )

Set style for line, col_start..col_end range if:

That line and column range exist in StyledBuffer

overwrite is true or existing style is Style::NoStyle or Style::Quotation

pub fn set_style( &mut self, line: usize, col: usize, style: Style, overwrite: bool )

Set style for line, col if:

That line and column exist in StyledBuffer

overwrite is true or existing style is Style::NoStyle or Style::Quotation