Trait rustc_errors::IntoDiagArg

pub trait IntoDiagArg {
    // Required method
    fn into_diag_arg(self) -> DiagArgValue;
}

Converts a value of a type into a DiagArg (typically a field of an Diag struct). Implemented as a custom trait rather than From so that it is implemented on the type being converted rather than on DiagArgValue, which enables types from other rustc_* crates to implement this.

Required Methods

fn into_diag_arg(self) -> DiagArgValue