Trait rustc_errors::IntoDiagnosticArg

pub trait IntoDiagnosticArg {
    // Required method
    fn into_diagnostic_arg(self) -> DiagnosticArgValue<'static>;
}

Converts a value of a type into a DiagnosticArg (typically a field of an IntoDiagnostic struct). Implemented as a custom trait rather than From so that it is implemented on the type being converted rather than on DiagnosticArgValue, which enables types from other rustc_* crates to implement this.

Required Methods

fn into_diagnostic_arg(self) -> DiagnosticArgValue<'static>