Trait rustc_errors::translation::Translate

pub trait Translate {
    // Required methods
    fn fluent_bundle(&self) -> Option<&Lrc<FluentBundle>>;
    fn fallback_fluent_bundle(&self) -> &FluentBundle;

    // Provided methods
    fn translate_messages(
        &self,
        messages: &[(DiagnosticMessage, Style)],
        args: &FluentArgs<'_>
    ) -> Cow<'_, str> { ... }
    fn translate_message<'a>(
        &'a self,
        message: &'a DiagnosticMessage,
        args: &'a FluentArgs<'_>
    ) -> Result<Cow<'_, str>, TranslateError<'_>> { ... }
}

Required Methods

fn fluent_bundle(&self) -> Option<&Lrc<FluentBundle>>

Return FluentBundle with localized diagnostics for the locale requested by the user. If no language was requested by the user then this will be None and fallback_fluent_bundle should be used.

fn fallback_fluent_bundle(&self) -> &FluentBundle

Return FluentBundle with localized diagnostics for the default locale of the compiler. Used when the user has not requested a specific language or when a localized diagnostic is unavailable for the requested locale.

Provided Methods

fn translate_messages( &self, messages: &[(DiagnosticMessage, Style)], args: &FluentArgs<'_> ) -> Cow<'_, str>

Convert DiagnosticMessages to a string, performing translation if necessary.

fn translate_message<'a>( &'a self, message: &'a DiagnosticMessage, args: &'a FluentArgs<'_> ) -> Result<Cow<'_, str>, TranslateError<'_>>

Convert a DiagnosticMessage to a string, performing translation if necessary.